taskbar.js: Define some 'public' constants using 'var'
This commit fixes the following JS warnings:

```
Some code accessed the property 'DASH_ITEM_HOVER_TIMEOUT' on the module 'taskbar'.
That property was defined with 'let' or 'const' inside the module. This was
previously supported, but is not correct according to the ES6 standard. Any
symbols to be exported from a module must be defined with 'var'. The property
access will work as previously for the time being, but please fix your code anyway.
```

```
Some code accessed the property 'DASH_ANIMATION_TIME' on the module 'taskbar'.
That property was defined with 'let' or 'const' inside the module. This was
previously supported, but is not correct according to the ES6 standard.
Any symbols to be exported from a module must be defined with 'var'.
The property access will work as previously for the time being, but please fix
your code anyway.
```

Signed-off-by: Alexander Rüedlinger <a.rueedlinger@gmail.com>