Implement animate-appicon-hover settings
The taskbar controls the animations by listening:
* appIcon 'notify::hover' and 'menu-state-changed' signals.
* scrollView 'motion-event' and 'leave-event' signals (ripple and plank animations).

The animations are processed in a new TaskbarItemContainer (that extends DashItemContainer).
Vertical animations are done with a Clutter clone because the appIcon actor cannot go outside the taskbar.
Horizontal animations (plank) are directly applied to the taskbarItemContainer.

AppIcon._onAnimateAppiconHoverChanged removes icon backgrounds when animations are enabled. It also increases the size of the icon texture in order to prevent it from being ugly when it is scaled.