Code cleanup in enable() and disable()
The current enable() and disable() methods in the panelManager
class do receive a 'reset' parameter to allow to "reset" and
recreate the panels without disconnecting and connecting again
signals and doing the whole disable-and-enable process.

This MR moves the common part into their own functions, and
thus removing the parameter and making the code simpler.