extension.js: Delay enablement till a monitor is available
When the extension is run with gnome-shell running in headless
mode, it tries to access monitor object upon enablement and crashes
thus it does not get enabled by the gnome-shell even after a monitor
becomes available.

This change delays the enablement till at least one monitor is
available on the system.