On startup, this._desktopIconsarea is not set as it is not intialized before this._setDesktopIconsMargins is called.
Therefore on DING and gtk4-ding, icons slide under the panel.

If the panel is repostioned or reset in any way, it behaves correctly.

With this change, desktopIconsUsable area is intialized in the constructor.

The margins are set to 0 in disable(reset = true);

The entire class is destroyed with disable();